set up kroki.io helper container, structurizr cli (B2KP-109) (#23)
Spin up a kroki.io container to render c4-plantuml and install the structurizr cli which is needed to generate c4-plantuml from structurizr dsl.

This wll be used with an update that will need to be made to `sphinxcontrib-kroki` that uses the structurizr cli to generate the c4-plantuml during documentation rendering so that a model/views defined across several files can be integrated into the sphinx documentation. 

Kroki does not accept multiple files, so the change above will be necessary to use it to render Structurizr DSL images into sphinx documentation.

The container doesn't exit reliably on Windows, which seems like it's a known issue https://github.com/microsoft/vscode-remote-release/issues/6056

The workaround is to manually stop the containers on your Windows machine if you need them stopped.